MatsubaraFrequency

A MatsubaraFrequency at temperature $T$ and with index $n$ can be generated using

T = 1.0
n = 5
v = MatsubaraFrequency(T, n, Fermion)
w = MatsubaraFrequency(T, n, Boson) 

Matsubara frequencies can be added, subtracted and their sign can be reversed, producing a new instance of MatsubaraFrequency.

v1 = v + v # type(v1) = :Boson
v2 = w - v # type(v2) = :Fermion
v3 = -v    # type(v3) = :Fermion 

Types

MatsubaraFunctions.AbstractParticle — Type

abstract type AbstractParticle

AbstractParticle type

MatsubaraFunctions.Fermion — Type

struct Fermion <: AbstractParticle

Fermionic particle type, used for MatsubaraFrequency and MatsubaraGrid constructors

MatsubaraFunctions.Boson — Type

struct Boson <: AbstractParticle

Bosonic particle type, used for MatsubaraFrequency and MatsubaraGrid constructors

MatsubaraFunctions.MatsubaraFrequency — Type

struct MatsubaraFrequency

MatsubaraFrequency type with fields:

T :: Float64 : temperature

val :: Float64 : position on the imaginary axis

idx :: Int64 : Matsubara index

type :: Symbol : particle type

Functions

MatsubaraFunctions.temperature — Function

function temperature(
    w :: MatsubaraFrequency
    ) :: Float64

Returns w.T

function temperature(
    grid :: MatsubaraGrid
    )    :: Float64

Returns grid.T

function temperature(
    f :: MatsubaraFunction{GD, SD, DD, Q},
    ) :: Float64 where {GD, SD, DD, Q <: Number}

Returns temperature for which f.grids are defined

MatsubaraFunctions.value — Function

function value(
    w :: MatsubaraFrequency
    ) :: Float64

Returns w.val

MatsubaraFunctions.index — Function

function index(
    w :: MatsubaraFrequency
    ) :: Int64

Returns w.idx

function index( 
    x :: MatsubaraIndex
    ) :: Int64

Returns x.idx

MatsubaraFunctions.type — Function

function type(
    w :: MatsubaraFrequency
    ) :: Symbol

Returns w.type

function type(
    grid :: MatsubaraGrid
    )    :: Symbol

Returns grid.type