MatsubaraFunction

A MatsubaraFunction is a collection of MatsubaraGrid instances together with an associated tensor structure $G_{i_1...i_n}$ for each point $(\omega_1, ..., \omega_m)$ in the cartesian product of the grids. Some possible constructors are

T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)

# 1D grid, rank 1 tensor with index dimension 1 (scalar valued)
f1_complex = MatsubaraFunction(g, 1) 
f1_real    = MatsubaraFunction(g, 1, Float64) 

# 1D grid, rank 1 tensor with index dimension 5 (vector valued)
f2_complex = MatsubaraFunction(g, 5) 
f2_real    = MatsubaraFunction(g, 5, Float64) 

# 1D grid, rank 2 tensor with index dimension 5 (matrix valued)
f3_complex = MatsubaraFunction(g, (5, 5)) 
f3_real    = MatsubaraFunction(g, (5, 5), Float64) 

# 2D grid, rank 2 tensor with index dimension 5 (matrix valued)
f4_complex = MatsubaraFunction((g, g), (5, 5)) 
f4_real    = MatsubaraFunction((g, g), (5, 5), Float64) 

There are two possible ways to access the data of a MatsubaraFunction, using either the bracket [] or the parenthesis () operator. The former can be used together with a set of linear indices or with a combination of MatsubaraFrequency objects and linear indices (for the tensor structure). It will return the value of the function precisely for the given arguments. () allows to substitute Float64 for the frequency arguments, in which case a multilinear interpolation is performed. In addition, () is well defined even for out of bounds access, using either a custom boundary condition or, for 1D grids, polynomial extrapolation.

ξ = 0.5
T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)
f = MatsubaraFunction(g, 1)

for v in g
    # if there is only one index of dimension 1, it does not need to be specified, i.e. 
    # f[v] can be used instead of f[v, 1] (also works for the '()' operator)
    f[v] = 1.0 / (im * value(v) - ξ)
end 

# access MatsubaraFunction data
v = g[rand(eachindex(g))]
println(f[v])        # fast data access, throws error if v is out of bounds
println(f(v))        # fast data access, defined even if v is out of bounds
println(f(value(v))) # slow data access, uses interpolation 

# fallback methods for out of bounds access
vp = MatsubaraFrequency(T, 256, Fermion)
println(f(vp))                                  # default x -> 0.0
println(f(vp; bc = x -> 1.0))                   # custom boundary condition x -> 1.0
println(f(vp; bc = x -> 1.0 / im / value(x)))   # custom boundary condition x -> 1.0 / im / value(x)
println(f(value(vp); bc = x -> 1.0 / im / x))   # custom boundary condition x -> 1.0 / im / x
println(f(vp; extrp = (true, ComplexF64(0.0)))) # polynomial extrapolation in 1D, constant term set to 0.0

MatsubaraFunction objects can be saved in HDF5 file format as

using MatsubaraFunctions 
using HDF5

file = h5open("test.h5", "w")
T    = 1.0
N    = 128
g    = MatsubaraGrid(T, N, Fermion)
f    = MatsubaraFunction(g, 1)

save_matsubara_function!(file, "func", f)
fp = load_matsubara_function(file, "func")
close(file)

Advanced Usage: Matsubara Sums

For MatsubaraFunction objects $G_{i_1 ... i_n}(i\omega)$ defined on 1D grids, we export the function sum_me, which computes the series $\Sigma_m G_{i_1 ... i_n}(i\omega_{m}) e^{i\omega_m 0^+}$ for $m \in \mathbb{Z}$ using tail fits of $G$ together with analytic formulas for summations of the form $\Sigma_m \frac{1}{(i\omega_m)^\alpha}e^{i\omega_m 0^+}$ with $\alpha \in \mathbb{N}$. This, however, requires $G$ to be representable by a Laurent series in an elongated annulus about the imaginary axis.

ξ = 0.5
T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)
f = MatsubaraFunction(g, 1)

for v in g
    f[v] = 1.0 / (im * value(v) - ξ)
end 

# evaluate the series and compare to analytic result
ρ(x, T) = 1.0 / (exp(x / T) + 1.0)
println(abs(sum_me(f, ComplexF64(0.0)) - (ρ(+ξ, T) - 1.0)))

Advanced Usage: Automated Symmetry Reduction

In many cases, the numerical effort of computing functions in the Matsubara domain can be drastically reduced by the use of symmetries. For one-particle Green's functions $G_{i_1 i_2}(i\omega)$, for example, hermicity of the Hamiltonian dictates that $G_{i_1 i_2}(i\omega) = G^{\star}_{i_2 i_1}(-i\omega)$, relating positive and negative Matsubara frequencies. We offer an automated way to compute the set of irreducible (i.e. unrelatable by symmetries) MatsubaraFunction components, as is illustrated in the following example

ξ = 0.5
T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)
f = MatsubaraFunction(g, 1)

for v in g
    f[v] = 1.0 / (im * value(v) - ξ)
end 

# complex conjugation acting on Green's function
function conj(
    w :: Tuple{MatsubaraFrequency},
    x :: Tuple{Int64}
    ) :: Tuple{Tuple{MatsubaraFrequency}, Tuple{Int64}, MatsubaraOperation}

    return (-w[1],), (x[1],), MatsubaraOperation(false, true)
end 

# compute the symmetry group 
SG = MatsubaraSymmetryGroup([MatsubaraSymmetry{1, 1}(conj)], f)

# obtain another Green's function by symmetrization
function init(
    w :: Tuple{MatsubaraFrequency},
    x :: Tuple{Int64}
    ) :: ComplexF64

    return f[w, x...]
end 

InitFunc = MatsubaraInitFunction{1, 1, ComplexF64}(init)
h        = MatsubaraFunction(g, 1)
SG(h, InitFunc)
@assert h == f

Advanced Usage: Running in parallel

To simplify the parallelization of algorithms when using the package, we export some useful methods based on the MPI.jl wrapper. For further information on how to set up MPI with Julia see https://github.com/JuliaParallel/MPI.jl.

using MatsubaraFunctions 
using MPI 

MPI.Init()
mpi_info()
mpi_println("I print on main.")
ismain = mpi_ismain() # ismain = true if rank is 0

T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)
f = MatsubaraFunction(g, 1)

# simple loop parallelization for UnitRange
for vidx in mpi_split(1 : length(g))
    println("My rank is $(mpi_rank()): $(vidx)")
end

# simple (+) allreduce
mpi_allreduce!(f)

In addition, calls of MatsubaraSymmetryGroup with an initialization function have an opt-in switch (mode) to enable parallel evaluation of the MatsubaraInitFunction (by default mode = :serial). If mode = :polyester, shared memory multithreading via the Polyester (https://github.com/JuliaSIMD/Polyester.jl) Julia package is used. This mode is recommended for initialization functions that are cheap to evaluate and are unlikely to benefit from Threads.@threads due to the overhead from invoking the Julia scheduler. For more expensive functions, users can choose between mode = :threads, which simply uses Threads.@threads, and mode = :hybrid. The latter combines both MPI and native Julia threads and can therefore be used to run calculations on multiple compute nodes.

Types

MatsubaraFunctions.MatsubaraFunction — Type

struct MatsubaraFunction{GD, SD, DD, Q <: Number}

MatsubaraFunction type with fields:

grids :: NTuple{GD, MatsubaraGrid} : collection of MatsubaraGrid

shape :: NTuple{SD, Int64} : shape of the tensor structure on every grid point

data :: Array{Q, DD} : data array

MatsubaraFunctions.MatsubaraOperation — Type

struct MatsubaraOperation

MatsubaraOperation type with fields:

sgn :: Bool : change sign?

con :: Bool : do complex conjugation?

MatsubaraFunctions.MatsubaraSymmetry — Type

struct MatsubaraSymmetry{GD, SD}

MatsubaraSymmetry type with fields:

f :: FunctionWrappers.FunctionWrapper{Tuple{NTuple{GD, MatsubaraFrequency}, NTuple{SD, Int64}, MatsubaraOperation}, Tuple{NTuple{GD, MatsubaraFrequency}, NTuple{SD, Int64}}}

MatsubaraSymmetry takes grid coordinates and tensor indices as input and returns a new set of coordinates and indices together with a MatsubaraOperation

MatsubaraFunctions.MatsubaraSymmetryGroup — Type

MatsubaraSymmetryGroup

MatsubaraSymmetryGroup type with fields:

classes :: Vector{Vector{Tuple{Int64, MatsubaraOperation}}} : list of collections of symmetry equivalent elements

speedup :: Float64 : expected speedup from the symmetry reduction

MatsubaraFunctions.MatsubaraInitFunction — Type

struct MatsubaraInitFunction{GD, SD, Q <: Number}

MatsubaraInitFunction type with fields:

f :: FunctionWrappers.FunctionWrapper{Q, Tuple{NTuple{GD, MatsubaraFrequency}, NTuple{SD, Int64}}}

MatsubaraInitFunction takes grid coordinates and tensor indices as input and returns value of type Q

Functions

MatsubaraFunctions.grids — Function

function grids(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: NTuple{GD, MatsubaraGrid} where {GD, SD, DD, Q <: Number}

Returns grids of f

function grids(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    idx :: Int64
    )   :: MatsubaraGrid where {GD, SD, DD, Q <: Number}

Returns f.grids[idx]

MatsubaraFunctions.grids_shape — Function

function grids_shape(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: NTuple{GD, Int64} where {GD, SD, DD, Q <: Number}

Returns length of grids

function grids_shape(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    idx :: Int64
    )   :: Int64 where {GD, SD, DD, Q <: Number}

Returns length of f.grids[idx]

MatsubaraFunctions.shape — Function

function shape(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: NTuple{SD, Int64} where {GD, SD, DD, Q <: Number}

Returns f.shape

function shape(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    idx :: Int64
    )   :: Int64 where {GD, SD, DD, Q <: Number}

Returns f.shape[idx]

MatsubaraFunctions.data_shape — Function

function data_shape(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: NTuple{DD, Int64} where {GD, SD, DD, Q <: Number}

Returns shape of f.data

function data_shape(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    idx :: Int64
    )   :: Int64 where {GD, SD, DD, Q <: Number}

Returns length of dimension idx of f.data

MatsubaraFunctions.absmax — Function

function absmax(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: Float64 where {GD, SD, DD, Q <: Number}

Returns largest element of f.data (in absolute terms)

Base.argmax — Function

function Base.:argmax(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: CartesianIndex{DD} where {GD, SD, DD, Q <: Number}

Returns position of largest element of f.data (in absolute terms)

MatsubaraFunctions.mpi_split — Function

function mpi_split(
    r :: UnitRange{Int64}
    ) :: UnitRange{Int64}

Splits UnitRange evenly among available MPI ranks (including main). Can, for example, be used to parallelize loops.

MatsubaraFunctions.mpi_allreduce! — Function

function mpi_allreduce!(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: Nothing where {GD, SD, DD, Q <: Number}

Inplace MPI reduction (+) for MatsubaraFunction

MatsubaraFunctions.mpi_ismain — Function

mpi_ismain() :: Bool

Returns true for MPI rank 0

MatsubaraFunctions.mpi_println — Function

function mpi_println(
    s :: String
    ) :: Nothing

Print string s on MPI rank 0

MatsubaraFunctions.mpi_info — Function

mpi_info() :: Nothing

Print information about available resources

MatsubaraFunctions.mpi_barrier — Function

mpi_barrier() :: Nothing

Place synchronization barrier for MPI ranks

MatsubaraFunctions.add — Function

function add(
    f1 :: MatsubaraFunction{GD, SD, DD, Q}, 
    f2 :: MatsubaraFunction{GD, SD, DD, Q}
    )  :: MatsubaraFunction{GD, SD, DD, Q} where {GD, SD, DD, Q <: Number}

Addition of two MatsubaraFunction, returns new MatsubaraFunction

MatsubaraFunctions.add! — Function

function add!(
    f1 :: MatsubaraFunction{GD, SD, DD, Q}, 
    f2 :: MatsubaraFunction{GD, SD, DD, Q}
    )  :: Nothing where {GD, SD, DD, Q <: Number}

Inplace addition of two MatsubaraFunction (f1 += f2)

MatsubaraFunctions.subtract — Function

function subtract(
    f1 :: MatsubaraFunction{GD, SD, DD, Q}, 
    f2 :: MatsubaraFunction{GD, SD, DD, Q}
    )  :: MatsubaraFunction{GD, SD, DD, Q} where {GD, SD, DD, Q <: Number}

Subtraction of two MatsubaraFunction, returns new MatsubaraFunction

MatsubaraFunctions.subtract! — Function

function subtract!(
    f1 :: MatsubaraFunction{GD, SD, DD, Q}, 
    f2 :: MatsubaraFunction{GD, SD, DD, Q}
    )  :: Nothing where {GD, SD, DD, Q <: Number}

Inplace subtraction of two MatsubaraFunction (f1 -= f2)

MatsubaraFunctions.mult — Function

function mult(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    val :: Qp
    )   :: MatsubaraFunction{GD, SD, DD, Q} where {GD, SD, DD, Q <: Number, Qp <: Number}

Multiplication of MatsubaraFunction with scalar, returns new MatsubaraFunction

MatsubaraFunctions.mult! — Function

function mult!(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    val :: Qp
    )   :: Nothing where {GD, SD, DD, Q <: Number, Qp <: Number}

Inplace multiplication of MatsubaraFunction with scalar (f *= val)

MatsubaraFunctions.set! — Function

function set!(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    val :: Qp,
    )   :: Nothing where {GD, SD, DD, Q <: Number, Qp <: Number}

Initialize MatsubaraFunction with val

function set!(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    arr :: Array{Qp, DD},
    )   :: Nothing where {GD, SD, DD, Q <: Number, Qp <: Number}

Initialize MatsubaraFunction with arr

function set!(
    f1 :: MatsubaraFunction{GD, SD, DD, Q},
    f2 :: MatsubaraFunction{GD, SD, DD, Q},
    )  :: Nothing where {GD, SD, DD, Q <: Number}

Initialize MatsubaraFunction with another MatsubaraFunction (f1 = f2)

MatsubaraFunctions.flatten — Function

function flatten(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: Vector{Q} where {GD, SD, DD, Q <: Number}

Flatten data array of MatsubaraFunction and return vector of the corresponding data type

MatsubaraFunctions.flatten! — Function

function flatten!(
    f :: MatsubaraFunction{GD, SD, DD, Q},
    x :: AbstractVector
    ) :: Nothing where {GD, SD, DD, Q <: Number}

Flatten data array of MatsubaraFunction into vector

MatsubaraFunctions.unflatten! — Function

function unflatten!(
    f :: MatsubaraFunction{GD, SD, DD, Q},
    x :: AbstractVector
    ) :: Nothing where {GD, SD, DD, Q <: Number}

Initialize data array of MatsubaraFunction from vector

MatsubaraFunctions.LinearIndex — Function

function LinearIndex(
    f :: MatsubaraFunction{GD, SD, DD, Q},
    w :: NTuple{GD, MatsubaraFrequency},
    x :: Vararg{Int64, SD} 
    ) :: Int64 where {GD, SD, DD, Q <: Number}

Returns linear index for access to f.data

function LinearIndex(
    f    :: MatsubaraFunction{GD, SD, DD, Q},
    cidx :: CartesianIndex{DD}
    )    :: Int64 where {GD, SD, DD, Q <: Number}

Returns linear index for access to f.data

function LinearIndex(
    f :: MatsubaraFunction{GD, SD, DD, Q},
    x :: Vararg{Int64, DD}
    ) :: Int64 where {GD, SD, DD, Q <: Number}

Returns linear index for access to f.data

MatsubaraFunctions.to_Matsubara — Function

function to_Matsubara(
    f    :: MatsubaraFunction{GD, SD, DD, Q},
    cidx :: CartesianIndex{DD}
    )    :: Tuple{NTuple{GD, MatsubaraFrequency}, NTuple{SD, Int64}} where {GD, SD, DD, Q <: Number}

Returns coordinates in grids and index of tensor structure

function to_Matsubara(
    f   :: MatsubaraFunction{GD, SD, DD, Q},
    idx :: Int64 
    )   :: Tuple{NTuple{GD, MatsubaraFrequency}, NTuple{SD, Int64}} where {GD, SD, DD, Q <: Number}

Returns coordinates in grids and index of tensor structure

MatsubaraFunctions.upper_tail_moments — Function

function upper_tail_moments(
    f  :: MatsubaraFunction{1, SD, DD, Q},
    α0 :: Q,
    x  :: Vararg{Int64, SD} 
    )  :: Tuple{Q, Q} where {SD, DD, Q <: Number}

Returns high frequency moments for quadratic model using upper grid bound. Here, α0 is the asymptotic limit for large positive frequencies. Note, that the distance between interpolation nodes is kept constant below T = 1.

MatsubaraFunctions.lower_tail_moments — Function

function lower_tail_moments(
    f  :: MatsubaraFunction{1, SD, DD, Q},
    α0 :: Q,
    x  :: Vararg{Int64, SD} 
    )  :: Tuple{Q, Q} where {SD, DD, Q <: Number}

Returns high frequency moments for quadratic model using lower grid bound. Here, α0 is the asymptotic limit for large negative frequencies. Note, that the distance between interpolation nodes is kept constant below T = 1.

MatsubaraFunctions.sum_me — Function

function sum_me(
    f  :: MatsubaraFunction{1, SD, DD, Q},
    α0 :: Q,
    x  :: Vararg{Int64, SD}
    )  :: Q where {SD, DD, Q <: Complex}

Computes the Matsubara sum (with regulator exp(-iw0+)) for a complex valued MatsubaraFunction on a 1D grid. Here, α0 is the asymptotic limit for large frequencies. This is only viable if f1 has a Laurent series representation with respect to an annulus about the imaginary axis.

function sum_me(
    f  :: MatsubaraFunction{1, 1, 2, Q},
    α0 :: Q
    )  :: Q where {Q <: Complex}

Computes the Matsubara sum (with regulator exp(-iw0+)) for a complex valued MatsubaraFunction on a 1D grid. Here, α0 is the asymptotic limit for large frequencies. This is only viable if f1 has a Laurent series representation with respect to an annulus about the imaginary axis. Requires shape(f, 1) == 1.

MatsubaraFunctions.density — Function

function density(
    f :: MatsubaraFunction{1, SD, DD, Q},
    x :: Vararg{Int64, SD}
    ) :: Q where {SD, DD, Q <: Complex}

Computes the density for a complex valued MatsubaraFunction on a 1D grid. Assumes that f decays to zero for large frequencies (as a single-particle Green's function would).

function density(
    f :: MatsubaraFunction{1, 1, 2, Q}
    ) :: Q where {Q <: Complex}

Computes the density for a complex valued MatsubaraFunction on a 1D grid. Assumes that f decays to zero for large frequencies (as a single-particle Green's function would). Requires shape(f, 1) == 1.

MatsubaraFunctions.sgn — Function

sgn(op :: MatsubaraOperation) :: Bool

Return op.sgn

MatsubaraFunctions.con — Function

con(op :: MatsubaraOperation) :: Bool

Return op.con

MatsubaraFunctions.save_matsubara_function! — Function

function save_matsubara_function!(
    h :: HDF5.File,
    l :: String,
    f :: MatsubaraFunction{Dg, Ds, Dt, Q}
    ) :: Nothing where {Dg, Ds, Dt, Q <: Number}

Save MatsubaraFunction f with label l to file h

MatsubaraFunctions.load_matsubara_function — Function

function load_matsubara_function(
    h :: HDF5.File,
    l :: String
    ) :: MatsubaraFunction

Load MatsubaraFunction with label l from file h

MatsubaraFunctions.save_matsubara_symmetry_group! — Function

function save_matsubara_symmetry_group!(
    h  :: HDF5.File,
    l  :: String,
    SG :: MatsubaraSymmetryGroup
    )  :: Nothing

Save MatsubaraSymmetryGroup `SG` with label `l` to file `h`

MatsubaraFunctions.load_matsubara_symmetry_group — Function

function load_matsubara_symmetry_group(
    h :: HDF5.File,
    l :: String
    ) :: MatsubaraSymmetryGroup

Load MatsubaraSymmetryGroup with label l from file h