MatsubaraGrid

A MatsubaraGrid is a sorted (symmetric) set of MatsubaraFrequency objects and can be constructed by

T  = 1.0
N  = 128
g1 = MatsubaraGrid(T, N, Fermion) # no. frequencies is 2N
g2 = MatsubaraGrid(T, N, Boson)   # no. frequencies is 2N - 1

MatsubaraGrid instances are iterable

T = 1.0
N = 128
g = MatsubaraGrid(T, N, Fermion)

for v in g
  println(value(v)) 
  println(index(v))
end

and can be evaluated using either a MatsubaraFrequency or Float64. As long as the input argument is in bounds, this will return the corresponding linear index of the grid in the former case and the linear index of the closest frequency in the latter case

T   = 1.0
N   = 128
g   = MatsubaraGrid(T, N, Fermion)
idx = rand(1 : length(g))
@assert g(g[idx]) == idx 
@assert g(value(g[idx])) == idx 

MatsubaraGrid objects can be saved in HDF5 file format as

using MatsubaraFunctions 
using HDF5

file = h5open("test.h5", "w")
T    = 1.0
N    = 128
g    = MatsubaraGrid(T, N, Fermion)

save_matsubara_grid!(file, "grid", g) 
gp = load_matsubara_grid(file, "grid")
close(file)

Types

MatsubaraFunctions.MatsubaraGrid — Type

struct MatsubaraGrid

MatsubaraGrid type with fields:

T :: Float64 : temperature

data :: Vector{MatsubaraFrequency} : list of MatsubaraFrequency objects

type :: Symbol : particle type

Functions

MatsubaraFunctions.first_index — Function

function first_index(
    grid :: MatsubaraGrid
    )    :: Int64

Returns the index of the first Matsubara frequency in grid

MatsubaraFunctions.last_index — Function

function last_index(
    grid :: MatsubaraGrid
    )    :: Int64

Returns the index of the last Matsubara frequency in grid

MatsubaraFunctions.index_range — Function

function index_range(
    grid :: MatsubaraGrid
    )    :: NTuple{2, Int64}

Returns indices of the first and last Matsubara frequency in grid

MatsubaraFunctions.is_inbounds — Function

function is_inbounds(
    w    :: MatsubaraFrequency,
    grid :: MatsubaraGrid
    )    :: Bool

Checks if w is contained in grid

function is_inbounds(
    w    :: Float64,
    grid :: MatsubaraGrid
    )    :: Bool

Checks if w lies within grid bounds

MatsubaraFunctions.N — Function

function N(
    grid :: MatsubaraGrid
    )    :: Int64

Returns that value N used to construct the grid

MatsubaraFunctions.first_value — Function

function first_value(
    grid :: MatsubaraGrid
    )    :: Float64

Returns the value of the first Matsubara frequency in grid

MatsubaraFunctions.last_value — Function

function last_value(
    grid :: MatsubaraGrid
    )    :: Float64

Returns the value of the last Matsubara frequency in grid

MatsubaraFunctions.value_range — Function

function value_range(
    grid :: MatsubaraGrid
    )    :: NTuple{2, Float64}

Returns values of the first and last Matsubara frequency in grid

MatsubaraFunctions.info — Function

function info(
    grid :: MatsubaraGrid
    )    :: Nothing

Prints summary of grid properties

function info(
    f :: MatsubaraFunction{GD, SD, DD, Q}
    ) :: Nothing where {GD, SD, DD, Q <: Number}

Prints summary of function properties

MatsubaraFunctions.save_matsubara_grid! — Function

function save_matsubara_grid!(
    h :: HDF5.File,
    l :: String,
    g :: MatsubaraGrid
    ) :: Nothing

Save MatsubaraGrid g with label l to file h

MatsubaraFunctions.load_matsubara_grid — Function

function load_matsubara_grid(
    h :: HDF5.File,
    l :: String
    ) :: MatsubaraGrid

Load MatsubaraGrid with label l from file h